[CWS] extract `PrepareK8SUserSessionContext` to own package
